Restructure tests folder.
