+++ .github/workflows/actions.yml
@@ -34,4 +34,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        pytest src/python/tensorflow_cloud/tests/core
+        pytest src/python/tensorflow_cloud/core/tests/unit/
