+++ src/python/tensorflow_cloud/core/tests/examples/call_run_on_notebook_with_keras_fit.py
