+++ src/python/tensorflow_cloud/core/tests/examples/call_run_on_script_with_keras_save_and_load.py
