+++ src/python/tensorflow_cloud/core/tests/examples/call_run_within_script_with_keras_fit.py
