+++ src/python/tensorflow_cloud/core/tests/testdata/keras_tuner_cifar_example.py
