+++ src/python/tensorflow_cloud/core/tests/testdata/mnist_example_using_fit_no_reqs.py
