+++ src/python/tensorflow_cloud/core/tests/testdata/save_and_load.py
