+++ src/python/tensorflow_cloud/core/tests/unit/containerize_test.py
@@ -32,7 +32,7 @@ except ImportError:
 class TestContainerize(unittest.TestCase):
     def setup(self):
         self.entry_point = (
-            "src/python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py"
+            "src/python/tensorflow_cloud/core/tests/testdata/mnist_example_using_fit.py"
         )
         self.chief_config = machine_config.COMMON_MACHINE_CONFIGS["K80_1X"]
         self.worker_config = machine_config.COMMON_MACHINE_CONFIGS["K80_1X"]
